TriggerCheckPermission from webui is been removed, to mandate checkpermission for all.

See merge request !376

+++ modules/CustomView/actions/Delete.php
@@ -10,10 +10,24 @@
 
 class CustomView_Delete_Action extends Vtiger_Action_Controller {
 
+	public function requiresPermission(\Vtiger_Request $request) {
+		$permissions = parent::requiresPermission($request);
+		$permissions[] = array('module_parameter' => 'sourceModule', 'action' => 'DetailView');
+		return $permissions;
+	}
+	
+	public function checkPermission(Vtiger_Request $request) {
+		return parent::checkPermission($request);
+	}
+	
 	public function process(Vtiger_Request $request) {
 		$customViewModel = CustomView_Record_Model::getInstanceById($request->get('record'));
 		$moduleModel = $customViewModel->getModule();
-
+		$customViewOwner = $customViewModel->getOwnerId();
+		$currentUser = Users_Record_Model::getCurrentUserModel();
+		if ((!$currentUser->isAdminUser()) || ($customViewOwner != $currentUser->getId())) {
+			throw new AppException(vtranslate('LBL_PERMISSION_DENIED'));
+		}
 		$customViewModel->delete();
 
 		$listViewUrl = $moduleModel->getListViewUrl();
